Rectangle input created.
